Fix a secondary check on geo-targeting
There is a bug related to ads targeting states/provinces
or metros which causes clicks and views to be undercounted.
When an ad is clicked or viewed, we double check the geo-targeting.
We already checked when the ad was first served but we check again.
This double check wasn't taking into account states or metros
and that check was failing.