Add initial display for current balances due
This is a stopgap until we add some kind of monthly balance display,
since that is the real data that we want for accounting and historical purposes.

Currently we don't track the timeframe that a payout is for,
which would make this code more explicit.
We could then also generate links to the ad report for that timeframe.