Fix a bug where existing AdType was excluded
This happened on the WTD ads I was testing with.
The AdType for the ad (RTD Sidebar) was attached to the RTD Publisher group,
but the WTD Publisher wasn't in it (so I could targets ads only to it).
This fixes it so that the Ad Type shows in this case,
but there's probably a better fix..