Define placement priority order
We didn't actually define the placement priority in our docs
other than to mention it. Nothing said what the highest priority
or lowest priority was.

Confusingly, the priorities were "reversed" with priority=1
being the highest. Firstly, I don't think anybody is using this.
RTD used to use this, but doesn't anymore.
The ad client doesn't support it.
We should consider removing or deprecating it.

However, I wanted to at least define it as I was making changes
to this code.