Publisher allowed domains
This is the first step toward restricting which domains
a publisher can show ads on. Right now, this just logs a message
at warning if this setting is turned on (off by default).

However, we have more and more publishers with 2 sites
and I'd like to start being a bit more systematic there.