Publisher payouts performance fix
This improves the performance of getting what a publisher is owed for
this month to date and for other periods (eg. publisher getting a payout
after a few months). This will also improve the performance of the
staff payout list view.

It uses the optimized aggregation we have for publisher paid
impressions. As a result, it can be up to 5 minutes out of date but
that's probably fine for almost all purposes.