User advertiser member check and constraint
I hit this extremely minor bug in production. If you try to add the same
user again as an authorized user on an advertiser, you hit a DB
constraint (500 error).

That's because the constraint added here exists
in production but not in the model. That is because at first the
UserAdvertiserMember model was just an automatic table and then we added
a new column to the table via migration.

This adds a secondary check to ensure we don't try to add the same user
again and adds the constraint in development. Hopefully there isn't an
issue adding that constraint again in prod or we could fake the
migration if desired since the constraint already exists.